style: Move TransitionProperty where it belongs.
Bug: 1419695
MozReview-Commit-ID: 9PN6VfbDbLA
